Extract field into its own file
